Deprecate depending on resolved configuration
In some edge case scenarios, it is possible to construct a graph where the configuration-being-resolved
is selected as a variant. Needing to support this edge case makes it very difficult to separate the
implementations of resolvable and consumable configurations.

We deprecate this behavior so that in Gradle 9+ we can more easily separate how resolvable and consumable
configurations are implemented.